Add conditional code to allow for checkpointing to the log file
This is a sample ... if acceptable, will expand it throughout the
code base, with the idea that it will hopefully be easier to isolate
the recent crashes to a specific function.

To enable, you would add -DCHECKPOINT_LOGGING to

Target -> FreeAPS -> Build Settings -> Swift Compiler - Custom Flags

   - Under "Other Swift Flags"

This could be extended to other areas of logging so that log.txt is
quieter unless there is debugging going on ( ie. get rid of all of the
DEV messages concerning OmniBLE for a regular build, shrinking log.txt
by almost 50%, and saving alot of calls to debug(), and writes to the
file system